fix: update example code to use refactored code
