+++ example_scripts/2_concept_generation.py
@@ -1,7 +1,8 @@
 from azure.ai.projects import AIProjectClient
-from atlas_rag import KnowledgeGraphGenerator, KnowledgeGraphExtractor, ProcessingConfig
+from atlas_rag.kg_construction.triple_extraction import KnowledgeGraphExtractor
+from atlas_rag.kg_construction.triple_config import ProcessingConfig
 from azure.identity import DefaultAzureCredential
-from atlas_rag.reader import LLMGenerator
+from atlas_rag.llm_generator import LLMGenerator
 from configparser import ConfigParser
 from argparse import ArgumentParser
 from openai import OpenAI
@@ -29,19 +30,20 @@ if __name__ == "__main__":
     )
     llm_generator = LLMGenerator(client=client, model_name=model_name)
     output_directory = f'import/{keyword}'
-    triple_generator = KnowledgeGraphGenerator(client, model_name=model_name,
-    max_new_tokens = 4096,
-    temperature = 0.1,
-    frequency_penalty = 1.1)
+    triple_generator = LLMGenerator(client, model_name=model_name)
     kg_extraction_config = ProcessingConfig(
         model_path=model_name,
-        data_directory="example_data",
+        data_directory="benchmark_data",
         filename_pattern=keyword,
-        batch_size=4,
+        batch_size_triple=4,
+        batch_size_concept=64,
         output_directory=f"{output_directory}",
-        record=True
+        current_shard_triple=args.slice,
+        total_shards_triple=args.total_slices,
+        record=True,
+        max_new_tokens=4096
     )
     kg_extractor = KnowledgeGraphExtractor(model=triple_generator, config=kg_extraction_config)
     kg_extractor.convert_json_to_csv()
-    kg_extractor.generate_concept_csv_temp(batch_size=64,language='en',record=True)
+    kg_extractor.generate_concept_csv_temp(language='en')
     kg_extractor.create_concept_csv()
